[MISC] Separate Bloom Filter strong types
Such that including the bloom_filter does not trigger deprecation warnings
